refactor: merge highscores.py into egg.py
